tts-miniservice/requirements.txt
# TTS Miniservice dependencies (Piper TTS)

piper-tts>=1.2.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
